fix printf format warnings

+++ PLC_esp8266/main/WiFi/WiFiService_Scanner.cpp
@@ -29,7 +29,7 @@ bool WiFiService::StartScan(const char *ssid,
 
     esp_err_t err = esp_wifi_scan_start(&scan_config, false);
     if (err != ESP_OK) {
-        ESP_LOGE(TAG_WiFiService_Scanner, "esp_wifi_scan_start err 0x%X", err);
+        ESP_LOGE(TAG_WiFiService_Scanner, "esp_wifi_scan_start err 0x%X", (unsigned int)err);
         return false;
     }
     return true;
@@ -51,7 +51,9 @@ int8_t WiFiService::Scanning(RequestItem *request,
         uint16_t number = sizeof(ap_info) / sizeof(ap_info[0]);
         err = esp_wifi_scan_get_ap_records(&number, ap_info);
         if (err != ESP_OK) {
-            ESP_LOGE(TAG_WiFiService_Scanner, "esp_wifi_scan_get_ap_records err 0x%X", err);
+            ESP_LOGE(TAG_WiFiService_Scanner,
+                     "esp_wifi_scan_get_ap_records err 0x%X",
+                     (unsigned int)err);
             break;
         }
 
@@ -79,7 +81,7 @@ int8_t WiFiService::Scanning(RequestItem *request,
         int64_t timespan = timeout_time - (uint64_t)esp_timer_get_time();
         ESP_LOGD(TAG_WiFiService_Scanner,
                  "process, uxBits:0x%08X, timespan:%lld",
-                 ulNotifiedValue,
+                 (unsigned int)ulNotifiedValue,
                  (long long)timespan);
 
         if (timespan <= 0) {
